+++ src/auth_utils/__init__.py
@@ -1,7 +1,7 @@
 from .utils import BaseUser, get_user, auth_required
 from .backends import JWTAuthBackend, APIKeyAuthBackend, AuthBackendsWrapper
 
-__version__ = "1.1.3"
+__version__ = "2.0.0"
 
 __all__ = [
     "BaseUser",